map & misc update
replaced walls and deleted floors. Added houses and alleyways. Replaced the farm with the apothecary, fully furnished. Partially furnished the inn. added a random meat spawn suitable for the inn. replaced puddle at apothecary with water pump. reduced the size of the plaza. added stone paths. changed the desc of mining lantern to flavor it better to a medieval setting as an oil lamp.  expanded the port.